feat: refactor OTP setup to use modal dialog
- Convert OtpSetupComponent to OtpSetupModal using Dialog layout
- Update SecurityMethodSelector to open OTP setup in modal overlay
- Maintain all existing functionality including error states and loading
- Follow same pattern as PinSetupModal for consistency
- Fix dialog open/close state management
